Create Movie Class - Static .fromString
Here we have our movie class and we have created 4 properties (contructors) - name, genre, rating, runtime. 
here final means the variable can not be changed once it is set in object construction. 
required means the name parameters must be provided with non-nullable value when calling constructors
Then we add our factory method .fromString - this creates a new movie object from a string input
Method takes our movie.String, which is a string representation of our movie object
because it is a list we can use index referencing. We have to use parse to turn the string to a  numeric data type representation

@override annotation is used to indicate that the toString method is overriding the default implementation of Object.toString() - this bit confuses me still.

Then we use the void main () function to provide the myMovie input and show how we want the output - then print the output.