Create data strucutres task - reversing a list
two examples of reversing lists and then also reversing a string.

First example we have a list but have no added a null safety in with it.

We use a void function (return type of method)
we call it reverseString (name our method)
tell the method what data we are manipulating - lists 
open our curley brackets to start
then we are just telling our method to calcuate the list length, then minus one from it. Then tell it to keep going until list is 0 or bigger; i -i.
so we are taking away final element then adding it to the start = reversing it.